requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
setuptools==68.2.2
wheel==0.41.2
pyOpenSSL==23.2.0
cryptography==41.0.3
waitress==2.1.2